issue 4226: move logic for update of run perf metrics from PodMonitor to NotificationAspect. To make sure, data is available when notification is about to send
